done with SnapshotMeta
